+++ frontend/src/component/project/Project/CreateProject/NewCreateProjectForm/ConfigButtons/TableConfigButton.tsx
@@ -1,0 +1,127 @@
+import { type FC, useState, useMemo } from 'react';
+import { CombinedSelect, type CombinedSelectProps } from './CombinedSelect';
+import type { DropdownListProps } from './DropdownList';
+import { TableSearchInput } from './SelectionButton.styles';
+import { InputAdornment } from '@mui/material';
+import Search from '@mui/icons-material/Search';
+import { ChangeRequestTable } from './ChangeRequestTable';
+
+type TableSelectProps = Pick<
+    CombinedSelectProps,
+    'button' | 'onOpen' | 'onClose' | 'description'
+> &
+    Pick<DropdownListProps, 'search'> & {
+        updateProjectChangeRequestConfiguration: {
+            disableChangeRequests: (env: string) => void;
+            enableChangeRequests: (
+                env: string,
+                requiredApprovals: number,
+            ) => void;
+        };
+        activeEnvironments: {
+            name: string;
+            type: string;
+        }[];
+        projectChangeRequestConfiguration: Record<
+            string,
+            { requiredApprovals: number }
+        >;
+    };
+
+export const TableSelect: FC<TableSelectProps> = ({
+    button,
+    search,
+    projectChangeRequestConfiguration,
+    updateProjectChangeRequestConfiguration,
+    activeEnvironments,
+    onOpen = () => {},
+    onClose = () => {},
+    ...props
+}) => {
+    const configured = useMemo(() => {
+        return Object.fromEntries(
+            Object.entries(projectChangeRequestConfiguration).map(
+                ([name, config]) => [
+                    name,
+                    { ...config, changeRequestEnabled: true },
+                ],
+            ),
+        );
+    }, [projectChangeRequestConfiguration]);
+
+    const tableEnvs = useMemo(
+        () =>
+            activeEnvironments.map(({ name, type }) => ({
+                name,
+                type,
+                ...(configured[name] ?? { changeRequestEnabled: false }),
+            })),
+        [configured, activeEnvironments],
+    );
+
+    const onEnable = (name: string, requiredApprovals: number) => {
+        updateProjectChangeRequestConfiguration.enableChangeRequests(
+            name,
+            requiredApprovals,
+        );
+    };
+
+    const onDisable = (name: string) => {
+        updateProjectChangeRequestConfiguration.disableChangeRequests(name);
+    };
+
+    const [anchorEl, setAnchorEl] = useState<HTMLDivElement | null>();
+    const [searchText, setSearchText] = useState('');
+
+    const filteredEnvs = tableEnvs.filter((env) =>
+        env.name.toLowerCase().includes(searchText.toLowerCase()),
+    );
+
+    const toggleTopItem = (event: React.KeyboardEvent) => {
+        if (
+            event.key === 'Enter' &&
+            searchText.trim().length > 0 &&
+            filteredEnvs.length > 0
+        ) {
+            const firstEnv = filteredEnvs[0];
+            if (firstEnv.name in configured) {
+                onDisable(firstEnv.name);
+            } else {
+                onEnable(firstEnv.name, 1);
+            }
+        }
+    };
+
+    return (
+        <CombinedSelect
+            button={button}
+            {...props}
+            anchorEl={anchorEl}
+            setAnchorEl={setAnchorEl}
+        >
+            <TableSearchInput
+                variant='outlined'
+                size='small'
+                value={searchText}
+                onChange={(event) => setSearchText(event.target.value)}
+                hideLabel
+                label={search.label}
+                placeholder={search.placeholder}
+                autoFocus
+                InputProps={{
+                    startAdornment: (
+                        <InputAdornment position='start'>
+                            <Search fontSize='small' />
+                        </InputAdornment>
+                    ),
+                }}
+                onKeyDown={toggleTopItem}
+            />
+            <ChangeRequestTable
+                environments={filteredEnvs}
+                enableEnvironment={onEnable}
+                disableEnvironment={onDisable}
+            />
+        </CombinedSelect>
+    );
+};
